rename source file inline_filter.rs to otf_filter.rs
